+++ frontend/src/router/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import stationRoutes from '@/router/modules/station'
-import lineRoutes from '@/router/modules/line'
+import lineRoutes from '@/router/modules/controller'
 import mainRoutes from '@/router/modules/main'
 import sectionRoutes from '@/router/modules/section'
 import mapRoutes from '@/router/modules/map'
