+++ frontend/src/store/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import station from '@/store/modules/station'
-import line from '@/store/modules/line'
+import line from '@/store/modules/controller'
 import snackbar from '@/store/modules/snackbar'
 import map from '@/store/modules/map'
 import path from '@/store/modules/path'
