+++ frontend/src/store/modules/line.js
@@ -1,6 +1,6 @@
 import { SET_LINE, SET_LINES } from '@/store/shared/mutationTypes'
 import { CREATE_LINE, DELETE_LINE, FETCH_LINES, EDIT_LINE, DELETE_SECTION, CREATE_SECTION, FETCH_LINE } from '@/store/shared/actionTypes'
-import LineService from '@/api/modules/line'
+import LineService from '@/api/modules/controller'
 
 const state = {
   line: {},
